Compare Classification Models

This function compares 13 different classification models using the caret package.

Usage

model_comparer(
  data,
  target_var,
  train_prop = 0.8,
  seed = 3456,
  for_utest = FALSE
)

Arguments

data

A data frame containing the dataset to be used for modeling.

target_var

The name of the target variable in the dataset.

train_prop

The proportion of data to be used for training (default is 0.8).

seed

The random seed for reproducibility (default is 3456).

for_utest

only for unit test when is TRUE (FALSE by default).

Value

A list containing the trained models and their performance metrics.

Examples

library(mlbench)
data("PimaIndiansDiabetes", package = "mlbench")
#results <- model_comparer(PimaIndiansDiabetes, "diabetes")